backend/requirements.txt
Flask==2.1.2
Flask-GraphQL==2.0.1
graphene==3.0.0b7
Flask-MongoEngine==1.0.0
Flask-JWT-Extended==4.3.1
pymongo==4.1.1
openai==0.10.2
langchain==0.1.0
dramatiq[rabbitmq, redis]==1.11.0
requests==2.27.1
python-dotenv==0.19.1
